Refactor forms components
