+++ packages/ui/src/components/forms/Label.tsx
@@ -1,0 +1,11 @@
+import styled from 'styled-components'
+import { Colors } from '../../constants'
+
+export const Label = styled.label`
+  margin-bottom: 4px;
+  font-size: 14px;
+  line-height: 20px;
+  font-weight: 700;
+  vertical-align: middle;
+  color: ${Colors.Black[900]};
+`
