+++ packages/ui/src/components/forms/index.tsx
@@ -1,0 +1,3 @@
+export * from './Input'
+export * from './Label'
+export * from './Switch'
