+++ packages/ui/src/modals/TransferModal/TransferDetailsModal.tsx
@@ -3,7 +3,7 @@ import React, { ReactElement, useState } from 'react'
 import styled from 'styled-components'
 import { AccountInfo } from '../../components/AccountInfo'
 import { ButtonPrimaryMedium, ButtonSecondarySmall } from '../../components/buttons'
-import { Label, NumberInput } from '../../components/Form'
+import { Label, NumberInput } from '../../components/forms'
 import { Modal, ModalBody, ModalFooter, ModalHeader } from '../../components/Modal'
 import { filterAccount, SelectAccount } from '../../components/selects/AccountSelectTemplate/SelectAccount'
 import { TokenValue } from '../../components/typography'
